Add: fill vacant position in database
If a slot, say slot 1, is free because its program exited, and the slots 2 and 5 are being used by other programs, allows for the slot 5 to switch to slot 1 to re-populate available space, in a cleaner way than leaving gaps.

Actual transformation of the window is not implemented yet, but the database manipulations on which it depends, is.