Rethink main window's database slot liberation
Reworked the way the main scripts exits. It now frees the slot using the main name inserted in it as an identifier, rather than using the slot number it was initially assigned during the first window repositioning. This allows for later repositioning of the window, which implies changing the slot assigned.