Implement denied script repositioning (early version)
If a script is denied, we want it to some special spot where they don't bother the reading of running script, and not on main screen too. For now, only one spot is defined for one window. If we want multiples we're going to need another table specifically for that.