Fix window repositioning order and add watcher cpu/fps display
Added a thread event to the shop watcher loop, so that it can signal to main when it starts. This allows for precise synch as to when "main" should give the order to "TWM" to manage the positioning of secondary windows who will just have spawned. Fps and CPU prints added to see script resources needs.