Implement socket communication between server and shop_watcher
First, this commit is a huge mess. Changed a lot of functions to be async in order for this to work, probably that the whole script is going to be async soon enough. There's now a server inside the shop_watcher module, listening to the main server. This way, when I send a ws request to the main server it can in turn communicate with the shop_watcher server to stop the loop or instruct more complex reactions, we'll see what we do with that. I BARELY UNDERSTAND HALF THE WRITTEN CODE CONCERNING THOSE DAMN SOCKET COMMUNICATIONS AND WILL REQUIRE QUITE A BIT OF TIME TO GRASP IT.