Implement stop subprocess execution from server
Made it so that now, when the server script receives a message from websockets asking for a subprocess termination, it will now connect as a client this time, with sockets, to the target subprocess server to send a "stop" message to it.

This allows for a StreamDeck-(websocket request)->Server-(socket message)->Subprocess communication chain. The server really serves as a launcher to start subprocesses, which would be impossible directly from the StreamDeck, which only has implementations to send WS request. (Unless I code a plugin for it, and GL with that)