feat: add PRIMES and Transmissions logic handling to GLN (Graph Logic Network)
Aight so PRIMES means the node connection is unlocked, but only until the next node is activated. This means, when a PRIMES connection is reached, either you activate that node with the next prompt, or you will to PRIME it again.

Next, for Transmissions: new third type of basic communication vector. This one is from system/script logic to the GLN, instead of between Prompter and GLN. Here, the new Expect. Failure transmission is called when the script detects that the EXPECTS relationship were not activated (basically meaning, the prompter failed to provide a matching answer to a question asked from the GLN).

Good stuff.