feat:wip: Implement GLN Initiation logic into code.
Hear me out, it's not good yet, and we're getting side tracked with some CYPHER building, but implemented still a decent way to handle how initiation should be: updated in a thread and checked frequently to see if we suddenly have to process a node after a certain time delay.

Have a good night.