refactor/feat: Implement REVERTS and DELAYS (unused rn) logic
New logic REVERTS which for now takes a target node, and removes all the impact that this target node's RELATIONSHIP had.. so, If I had a node un/locking 3 others, targeting the root node with REVERT will remove the un/lock from all 3 others. Very useful to keep the graph readable, avoiding having to reconnect a response to all  the nodes.

The DELAYS logic is for setting the countdown of initially an initiation back to a different value. But I ended up kind of implementing it to really be able to change any time-bound item. Although for now, i just set ALL time-bound items of a node to the set duration. Ended up not using it for now (replaced with a simple INITIATES fitting the use case) but it's there and can developed further.

Also refactor for the modifying relationship (so right now: REVERTS and DISABLES) to have their own function.