feat:add LogicGate Nodes ! With conditional relationship
Yeaaah that's pretty cool uh. They are not super duper deep implemented but they are fairly robust and are are to check if any attribute in conversation state corresponds with a target node. Basically able to tell wether a target or multiple target node are locked, unlocked, expected... etc and activate another node if the conditions are met. Could be further developed but for now this will do, at some point I have to stop adding feature to the GLN and start using it to give a voice to robeau :).