feat: Add Greeting AND "Greeting + Immediate Prompt" recognition
Ok so we will parse the first few words uttered so as to know wether we just greeted robeau OR greeted AND immediately, within the same sentence, passed a prompt. WHICH IS VERY COOL.

That and we add the proper GLN logic + DB nodes to handle the basic of the greeting stage.