feat: Add GLN thread pausing while speaking in mic + GLN prompt recon
Allows to pause the update thread running in GLN while we speak into the mic as to avoid overlap between our voice and a sudden node trigger that had been initiated before hand. This does not extend the duration of time bound items, since their duration is based on a `start_time` vs `actual_time` delta. Additionally, we make GLN process the logic of recognizing a prompt has been uttered, and if it wasn't recognized at all, to provide an answer for that.

#TODO: make the greetings stages of robeau more GLN centric with output/input node rather than depends on the code of the robeau module.