feat: Add context handling for stubborn state
Specifically, how to handle the fact that in both normal and stubborn states I might want to utter the same whisper and we will have to not trigger them both at once. Pretty tricky.

Got an approach with text mutation working but I think it'll be better with a Label based approach. Will do next and compare.